acertar o parse Integer.valueOf(n), intValue(n), BigInteger.valueOf(n)

+++ src/main/java/com/server/soap/webservices/customersadministration/bean/Customer.java
@@ -4,13 +4,12 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import java.math.BigInteger;
 
 @Entity
 public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private BigInteger id;
+    private Integer id;
 
     private String name;
     private String phone;
@@ -19,18 +18,18 @@ public class Customer {
     public Customer() {
     }
 
-    public Customer(BigInteger id, String name, String phone, String email) {
+    public Customer(Integer id, String name, String phone, String email) {
         this.id = id;
         this.name = name;
         this.phone = phone;
         this.email = email;
     }
 
-    public BigInteger getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(BigInteger id) {
+    public void setId(Integer id) {
         this.id = id;
     }
